Update README.md (#1)
This was a great revised draft! I've made minor edits for clarity/grammar and also attempted to add more transitions and explanation in the step by step process to make it easier to follow. I do have a few questions in brackets for you to address as well. Thank you!