Copy edit (#4)
Description of changes:

This is a copy edit to prep this article for publication on the AWS Open Source Blog. I've removed capital letters from endpoint model, according to the service names style guide. I've spelled out all numbers under 10, according to AP style. And changed setup to set up when it's used as a verb. There are a few typos I fixed as well.